stored all country names locally
