+++ shared/constants.ts
@@ -1,0 +1,3 @@
+import countryNames from "../data/User/countryNames.json";
+
+export const COUNTRY_NAMES = new Set(countryNames)
